requirements.txt
ffmpeg_python==0.2.0
gliclass==0.1.11
gliner==0.2.21
immanuel==1.5.0
loguru==0.7.2
matplotlib==3.8.4
mutagen==1.47.0
numpy==2.3.3
pandas==2.3.3
pymediainfo==7.0.1
Requests==2.32.5
rich==14.1.0
scikit_learn==1.4.2
spacy==3.8.7
stylecloud==0.5.2
toml==0.10.2
torch==2.7.1+cu128
tqdm==4.67.1
transformers==4.46.3
wordcloud==1.9.4
